refactor(run): move setupSSHAgent to its own file with tests
Relocate the extracted SSH-agent setup (setupSSHAgent + sshAgentSetup) out
of manager_create.go into manager_ssh.go, matching the package's
purpose-named-file convention (manager_lifecycle.go, manager_monitor.go,
...). Pure move, no logic change.

Add manager_ssh_test.go covering the guard branches the extraction made
unit-testable: no grants, missing SSH_AUTH_SOCK, credential-store error, no
SSH key mappings, and upstream-agent connect failure. These exercise the
early-return paths with a faked openCredStore closure and a real temp
FileStore, no live SSH agent required. The success paths (server
start + env/mount wiring) still need a real/faked agent and remain covered
only by e2e.
